+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 ### Features:
 - Added color support for double hook messages (only when not in a party)
+- Added {total} variable to Rare SC messages
 
 ### Fixes:
 - Fixed double hook messages not being sent in party
+- Fixed Rare SC messages {count} being the total instead of count since
 
 ### Back-end:
 - Nothing